requirements.txt
opencv-python
numpy
RPi.GPIO
onnxruntime
PyYAML
loguru
pyserial
